+++ app/src/main/java/com/comjeong/nomadworker/domain/repository/home/HomeRepository.kt
@@ -2,7 +2,6 @@ package com.comjeong.nomadworker.domain.repository.home
 
 import com.comjeong.nomadworker.data.model.home.UpdateCurrentLocationRequestData
 import com.comjeong.nomadworker.domain.model.home.LocationCategoryResult
-import com.comjeong.nomadworker.domain.model.home.LocationPlaceResult
 import com.comjeong.nomadworker.domain.model.home.UpdateCurrentLocationResult
 
 interface HomeRepository {
