+++ app/src/main/java/com/comjeong/nomadworker/domain/repository/place/PlaceRegionRepository.kt
@@ -1,6 +1,6 @@
 package com.comjeong.nomadworker.domain.repository.place
 
-import com.comjeong.nomadworker.domain.model.home.LocationPlaceResult
+import com.comjeong.nomadworker.domain.model.place.LocationPlaceResult
 
 interface PlaceRegionRepository {
     suspend fun getPlaceByLocationName(categoryName: String): LocationPlaceResult
