+++ app/src/main/res/layout/item_search_place.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="place"
-            type="com.comjeong.nomadworker.domain.model.home.LocationPlaceResult.Result.Place" />
+            type="com.comjeong.nomadworker.domain.model.place.LocationPlaceResult.Result.Place" />
 
     </data>
